FastAPIBackend/lawyerai-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
jinja2
gunicorn
python-multipart
aiosmtplib
email-validator
python-dotenv
